Night Time Farming
An entire new Collection of Features has been partially implemented. A Play Test has shown that having an Exploration Phase at the Start of the Mod is stressful after doing the early Game Exploration. Therefore, a new Farming and Crafting Progression System is added to the Start of the Mod. So far, only the Plants are functional and Crafting Blocks still need to be added. The Theme of the Farming Mechanic is that the new Plants only grow at Night.